requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.2
XlsxWriter==3.2.0
seatable-api==2.8.1
openpyxl==3.1.5
python-dotenv==1.0.1